// backend/src/types/args.type.ts
export interface Args {
  host?: string | undefined;
  port?: string | undefined;
  dbname?: string | undefined;
  user?: string | undefined;
  password?: string | undefined;
  url?: string | undefined;
  "listen-port"?: string | undefined;
  version?: boolean | undefined;
  help?: boolean | undefined;
}